+++ CHANGELOG.md
@@ -295,3 +295,4 @@
 * OnFire and OnReload scripts are now passed item model in addition to instance
 * Added PlaySound overloads with more params (like PlaySoundEx, but does not throw)
 * Added options to not receive messages on inactive components (QdmsMessageInterface and QdmsMessageComponent)
+* QdmsMessageComponent can now translate and rebroadcast as Unity messages [Experimental]